rename/자랑,추천,질문 게시판 이름 수정
