--- src/components/QnABoard.js
@@ -1,12 +1,0 @@
-// 고민게시판
-import React from 'react';
-
-const QnABoard = () => {
-  return (
-    <div>
-      <h1>고민 해결소</h1>
-    </div>
-  );
-};
-
-export default QnABoard;
