+++ src/components/QuestionsBoard.js
@@ -1,0 +1,12 @@
+// 고민게시판
+import React from 'react';
+
+const QuestionsBoard= () => {
+  return (
+    <div>
+      <h1>고민 게시판</h1>
+    </div>
+  );
+};
+
+export default QuestionsBoard;
